third_party/WebKit/LayoutTests/usb/mock-services.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/device.mojom.js"></script>
<script src="resources/device_manager.mojom.js"></script>
<script src="resources/permission_provider.mojom.js"></script>
<script src="resources/fake-devices.js"></script>
<script src="resources/usb-helpers.js"></script>
<script>
'use strict';

usb_test(usb => {
  assert_true(usb instanceof Object);
  assert_true(usb.DeviceManager instanceof Object);
  assert_true(usb.Device instanceof Object);
  assert_true(usb.PermissionProvider instanceof Object);
  assert_true(usb.mockDeviceManager instanceof Object);
}, 'USB Mojo bindings and mock interfaces are available to tests.')
</script>
